Deliver xml file at the publication to DG #130
